feat: added the last components
sliced the components and added the tabs navigation
